Refactor to {{search-filter}} component
